feat: move sidebar to layout
