+++ src/components/common/sidebar/__mocks__/Sidebar.tsx
@@ -1,0 +1,7 @@
+import { PropsWithChildren } from 'react';
+
+function Sidebar({ children }: PropsWithChildren) {
+  return <div>{children}</div>;
+}
+
+export default Sidebar;
