+++ src/pages/index.test.tsx
@@ -2,6 +2,8 @@ import { render } from '@testing-library/react';
 
 import HomePage from './index.page';
 
+jest.mock('@/components/common/sidebar/Sidebar');
+
 describe('HomePage', () => {
   const renderHomePage = () => render(
     <HomePage />,
